Move pkg/cache to internal
