+++ internal/storage/v1/elasticsearch/spanstore/service_operation.go
@@ -15,8 +15,8 @@ import (
 	"github.com/olivere/elastic"
 	"go.uber.org/zap"
 
+	"github.com/jaegertracing/jaeger/internal/cache"
 	"github.com/jaegertracing/jaeger/internal/storage/v1/elasticsearch/spanstore/internal/dbmodel"
-	"github.com/jaegertracing/jaeger/pkg/cache"
 	"github.com/jaegertracing/jaeger/pkg/es"
 )
